If there aren't any `_onStarting` events; generally the whole startup can bypass the `async`.

Using `ReferenceEquals` on `TaskCache.CompletedTask` rather than doing the bit manipulations that Task does to work out if it has completed sync, as is simpler. 

Better bits from https://github.com/aspnet/KestrelHttpServer/pull/1477