@benaadams 2017-04-08
Linux issue https://github.com/aspnet/KestrelHttpServer/issues/1632
> Failed   ConnectionClosesWhenFinReceivedBeforeRequestCompletes(listenOptions: http://127.0.0.1:0)
Error Message:
 Assert.Equal() Failure
          ↓ (pos 0)
Expected: 
Actual:   HTTP/1.1 400 Bad Request\r\nConnection: clo···
          ↑ (pos 0)