@benaadams 2017-04-08
Cuts out two functions from the call chain

Pre
![Pre](https://aoa.blob.core.windows.net/aspnet/async-2.png)

Post
![Post](https://aoa.blob.core.windows.net/aspnet/async-1.png)

Trying to find out why `WriteAsync` allocates 64 bytes in but `Write` doesn't