@JamesNK 2017-04-08
@davidfowl I found with Json.NET async work that async methods are a lot slower to call than normal methods. The async state machine adds a lot of overhead.