@davidfowl 2017-04-08
@JamesNK I'm well aware, I'm still questioning the gains. We make tons of micro optimizations in kestrel that usually add up but there is a cost of maintaining hard to read code. So there's a point at which the optimization has to be worth the ugly and we have micro benchmarks checked in for jus that reason.