@benaadams 2017-04-08
64B allocation is just a measuring artefact; issue raised https://github.com/dotnet/BenchmarkDotNet/issues/415