@benaadams 2017-04-08
`WriteAsync` checks `_onStarting` under lock, and `Write` checks it not under lock. Should either be both do or neither. Changing it to both