@benaadams 2017-04-09
Complete Results
```
     Method | WithHeaders | Chunked | OnStarting |        Mean |    StdErr |       Op/s | Allocated |
----------- |------------ |-------- |----------- |------------ |---------- |----------- |---------- |
 WriteAsync |       False |   False |      Async | 143.7179 ns | 1.3632 ns | 6958076.98 |      78 B |
 WriteAsync |       False |   False |       None | 142.8760 ns | 0.4573 ns | 6999077.81 |      78 B |
 WriteAsync |       False |   False |       Sync | 141.6665 ns | 0.1538 ns | 7058832.93 |      78 B |
 WriteAsync |       False |    True |      Async | 140.5505 ns | 0.1393 ns | 7114879.17 |      78 B |
 WriteAsync |       False |    True |       None | 141.8294 ns | 0.3130 ns | 7050726.11 |      78 B |
 WriteAsync |       False |    True |       Sync | 142.3879 ns | 0.3813 ns | 7023069.85 |      78 B |
 WriteAsync |        True |   False |      Async | 657.5599 ns | 0.7676 ns | 1520773.97 |     378 B |
 WriteAsync |        True |   False |       None | 422.1309 ns | 1.9038 ns | 2368933.23 |     178 B |
 WriteAsync |        True |   False |       Sync | 692.7274 ns | 7.1783 ns | 1443569.30 |     306 B |
 WriteAsync |        True |    True |      Async | 643.7326 ns | 1.0632 ns | 1553440.08 |     378 B |
 WriteAsync |        True |    True |       None | 420.6879 ns | 0.2489 ns | 2377059.10 |     178 B |
 WriteAsync |        True |    True |       Sync | 627.1672 ns | 0.7878 ns | 1594471.05 |     306 B |
```

Before and after of the more interesting portion

Before
```
     Method | WithHeaders | Chunked | OnStarting |        Mean |    StdErr |       Op/s | Allocated |
----------- |------------ |-------- |----------- |------------ |---------- |----------- |---------- |
 WriteAsync |        True |   False |      Async | 667.5727 ns | 1.3135 ns | 1497964.30 |     378 B |
 WriteAsync |        True |   False |       None | 529.8076 ns | 0.6551 ns | 1887477.81 |     178 B |
 WriteAsync |        True |   False |       Sync | 657.7463 ns | 2.3489 ns | 1520343.12 |     306 B |
```

After
```
     Method | WithHeaders | Chunked | OnStarting |        Mean |    StdErr |       Op/s | Allocated |
----------- |------------ |-------- |----------- |------------ |---------- |----------- |---------- |
 WriteAsync |        True |   False |      Async | 657.5599 ns | 0.7676 ns | 1520773.97 |     378 B |
 WriteAsync |        True |   False |       None | 422.1309 ns | 1.9038 ns | 2368933.23 |     178 B |
 WriteAsync |        True |   False |       Sync | 692.7274 ns | 7.1783 ns | 1443569.30 |     306 B |
```