@benaadams 2017-04-09
Hmm... sync and async bit weird; will try local copy of completed task