@benaadams 2017-04-09
(Updated) Before and after of the more interesting portion

Before (https://github.com/benaadams/KestrelHttpServer/tree/testwriting for exposed methods)
```
     Method | WithHeaders | Chunked | OnStarting |        Mean |    StdDev |       Op/s | Allocated |
----------- |------------ |-------- |----------- |------------ |---------- |----------- |---------- |
 WriteAsync |        True |   False |       None | 525.3814 ns | 2.0567 ns | 1903379.13 |     178 B |
 WriteAsync |        True |   False |       Sync | 637.3575 ns | 3.8968 ns | 1568978.12 |     306 B |
 WriteAsync |        True |   False |      Async | 636.1822 ns | 4.5565 ns | 1571876.73 |     306 B |
```

After (this PR)
```
     Method | WithHeaders | Chunked | OnStarting |        Mean |    StdDev |       Op/s | Allocated |
----------- |------------ |-------- |----------- |------------ |---------- |----------- |---------- |
 WriteAsync |        True |   False |       None | 421.0196 ns | 7.0106 ns | 2375186.07 |     178 B |
 WriteAsync |        True |   False |       Sync | 543.3438 ns | 4.4737 ns | 1840455.51 |     306 B |
 WriteAsync |        True |   False |      Async | 658.5292 ns | 3.4424 ns | 1518535.49 |     306 B |
```

/cc @davidfowl da numbers 😄