@benaadams 2017-04-09
Added tests as PR https://github.com/aspnet/KestrelHttpServer/pull/1642
```
     Method | WithHeaders | Chunked | OnStarting |        Mean |    StdDev |      Median |       Op/s | Allocated |
----------- |------------ |-------- |----------- |------------ |---------- |------------ |----------- |---------- |
 WriteAsync |       False |   False |      Async | 141.0204 ns | 0.7899 ns | 140.9354 ns | 7091171.73 |      78 B |
 WriteAsync |       False |   False |       None | 140.9476 ns | 0.7818 ns | 141.0736 ns | 7094835.66 |      78 B |
 WriteAsync |       False |   False |       Sync | 141.6527 ns | 1.5219 ns | 141.5544 ns | 7059519.97 |      78 B |
 WriteAsync |       False |    True |      Async | 140.4693 ns | 0.7723 ns | 140.4000 ns | 7118994.26 |      78 B |
 WriteAsync |       False |    True |       None | 140.1714 ns | 1.0770 ns | 140.0190 ns | 7134121.96 |      78 B |
 WriteAsync |       False |    True |       Sync | 142.6427 ns | 3.2782 ns | 141.0240 ns | 7010522.14 |      78 B |
 WriteAsync |        True |   False |      Async | 625.4775 ns | 4.3903 ns | 625.2574 ns | 1598778.59 |     306 B |
 WriteAsync |        True |   False |       None | 417.3450 ns | 2.5388 ns | 416.4601 ns | 2396098.98 |     178 B |
 WriteAsync |        True |   False |       Sync | 520.6009 ns | 4.0205 ns | 519.5717 ns | 1920857.06 |     306 B |
 WriteAsync |        True |    True |      Async | 624.6187 ns | 5.1475 ns | 624.9212 ns | 1600976.81 |     306 B |
 WriteAsync |        True |    True |       None | 420.0299 ns | 1.5173 ns | 419.6871 ns | 2380782.76 |     178 B |
 WriteAsync |        True |    True |       Sync | 522.6808 ns | 4.8042 ns | 522.3586 ns | 1913213.42 |     306 B |
```