@benaadams 2017-04-10
Re-based and reopened as https://github.com/aspnet/KestrelHttpServer/pull/1644