@benaadams 2017-04-10
Rebased, removed benchmark changes (mostly), reduced code churn