@halter73 2017-08-18
I rebased your changes and here are the current before and after numbers on my laptop given the updated micro benchmarks from #1642.

### Old dev:

```
     Method | WithHeaders | Chunked | OnStarting |          Mean |     StdErr |      StdDev |        Median |       Op/s | Allocated |
----------- |------------ |-------- |----------- |-------------- |----------- |------------ |-------------- |----------- |---------- |
 WriteAsync |       False |   False |      Async |   223.1018 ns |  3.8948 ns |  21.3326 ns |   211.0020 ns | 4482258.18 |      78 B |
 WriteAsync |       False |   False |       None |   209.1302 ns |  1.8205 ns |   9.9715 ns |   206.4410 ns | 4781711.13 |      78 B |
 WriteAsync |       False |   False |       Sync |   207.6673 ns |  1.0871 ns |   5.9542 ns |   205.9419 ns | 4815395.51 |      78 B |
 WriteAsync |       False |    True |      Async |   221.9945 ns |  4.4913 ns |  24.5997 ns |   210.8805 ns |    4504616 |      78 B |
 WriteAsync |       False |    True |       None |   221.8296 ns |  3.4974 ns |  19.1559 ns |   215.2639 ns | 4507964.57 |      78 B |
 WriteAsync |       False |    True |       Sync |   217.1743 ns |  2.8331 ns |  15.5176 ns |   209.6381 ns | 4604596.22 |      78 B |
 WriteAsync |        True |   False |      Async | 1,081.5509 ns | 31.7755 ns | 174.0416 ns | 1,134.3591 ns |  924598.21 |     305 B |
 WriteAsync |        True |   False |       None |   685.4649 ns |  6.2189 ns |  34.0625 ns |   673.7863 ns |  1458863.8 |     177 B |
 WriteAsync |        True |   False |       Sync |   891.8900 ns | 15.1367 ns |  82.9071 ns |   857.5039 ns | 1121214.47 |     305 B |
 WriteAsync |        True |    True |      Async |   834.2390 ns |  8.3393 ns |  45.6765 ns |   832.9739 ns | 1198697.28 |     305 B |
 WriteAsync |        True |    True |       None |   720.8754 ns | 15.8111 ns |  86.6010 ns |   688.5577 ns | 1387202.26 |     177 B |
 WriteAsync |        True |    True |       Sync |   818.8223 ns |  3.8527 ns |  21.1021 ns |   813.6414 ns | 1221266.18 |     305 B |
```

### This PR

```
     Method | WithHeaders | Chunked | OnStarting |        Mean |     StdErr |      StdDev |      Median |       Op/s | Allocated |
----------- |------------ |-------- |----------- |------------ |----------- |------------ |------------ |----------- |---------- |
 WriteAsync |       False |   False |      Async | 227.7008 ns |  7.5025 ns |  41.0929 ns | 202.3973 ns |  4391727.9 |      78 B |
 WriteAsync |       False |   False |       None | 197.3162 ns |  0.6649 ns |   3.6417 ns | 196.8270 ns | 5068006.96 |      78 B |
 WriteAsync |       False |   False |       Sync | 196.7176 ns |  0.5358 ns |   2.9346 ns | 196.6478 ns | 5083428.64 |      78 B |
 WriteAsync |       False |    True |      Async | 202.0139 ns |  0.8840 ns |   4.8418 ns | 201.1500 ns | 4950155.12 |      78 B |
 WriteAsync |       False |    True |       None | 201.4103 ns |  0.5774 ns |   3.1624 ns | 201.3188 ns | 4964988.43 |      78 B |
 WriteAsync |       False |    True |       Sync | 201.8038 ns |  0.6568 ns |   3.5976 ns | 201.0586 ns |  4955307.6 |      78 B |
 WriteAsync |        True |   False |      Async | 830.9345 ns | 19.4620 ns | 106.5979 ns | 781.5924 ns |  1203464.3 |     305 B |
 WriteAsync |        True |   False |       None | 543.3715 ns |  5.2474 ns |  28.7412 ns | 530.7326 ns | 1840361.55 |     177 B |
 WriteAsync |        True |   False |       Sync | 697.5552 ns | 11.1772 ns |  61.2198 ns | 676.5557 ns | 1433578.28 |     305 B |
 WriteAsync |        True |    True |      Async | 916.9603 ns | 35.8063 ns | 196.1193 ns | 802.1534 ns |  1090559.8 |     305 B |
 WriteAsync |        True |    True |       None | 714.5119 ns | 32.0643 ns | 175.6234 ns | 653.1019 ns | 1399556.89 |     177 B |
 WriteAsync |        True |    True |       Sync | 799.3473 ns | 38.1445 ns | 208.9260 ns | 658.3018 ns | 1251020.67 |     305 B |
```

Seems better overall to me, and all the tests still pass. I reviewed this pretty thoroughly while rebasing and it seems reasonable, so I'm merging.